signin jwt token verification

+++ Backend/routes/cart.js
@@ -10,7 +10,7 @@ module.exports = router;
 
 //Get Cart Contents
 router.get('/cart', (request, response) => {
-	const statement = `SELECT cart_id, user_id, prod_id, cart_quantity FROM cart`;
+	const statement = `SELECT cart_id, user_id, prod_id, cart_quantity FROM cart where user_id = ${request.id}`;
 	db.execute(statement, (error, data) => {
 		response.send(utils.createResult(error, data));
 	});
@@ -18,8 +18,8 @@ router.get('/cart', (request, response) => {
 
 //Insert new content in Cart
 router.post('/cart', (request, response) => {
-	const { user_id, prod_id, cart_quantity } = request.body;
-	const statement = `INSERT INTO cart(user_id, prod_id, cart_quantity) VALUES('${user_id}', '${prod_id}', '${cart_quantity}') `;
+	const { prod_id, cart_quantity } = request.body;
+	const statement = `INSERT INTO cart(user_id, prod_id, cart_quantity) VALUES('${request.id}', '${prod_id}', '${cart_quantity}') `;
 	db.execute(statement, (error, data) => {
 		response.send(utils.createResult(error, data));
 	});
@@ -29,7 +29,7 @@ router.post('/cart', (request, response) => {
 router.put('/cart/:cart_id', (request, response) => {
 	const { cart_quantity } = request.body;
 	const { cart_id } = request.params;
-	const statement = `UPDATE cart SET cart_quantity = '${cart_quantity}' WHERE cart_id = '${cart_id}'`;
+	const statement = `UPDATE cart SET cart_quantity = '${cart_quantity}' WHERE cart_id = '${cart_id}' and user_id=${request.id}`;
 	db.execute(statement, (error, data) => {
 		response.send(utils.createResult(error, data));
 	});
@@ -38,7 +38,7 @@ router.put('/cart/:cart_id', (request, response) => {
 // Delete Cart
 router.delete('/cart/:cart_id', (request, response) => {
 	const { cart_id } = request.params;
-	const statement = `DELETE FROM cart WHERE cart_id = '${cart_id}'`;
+	const statement = `DELETE FROM cart WHERE cart_id = '${cart_id}' and user_id =${request.id} `;
 	db.execute(statement, (error, data) => {
 		response.send(utils.createResult(error, data));
 	});
